removed useGroupMenu hook, improved route navigation in GroupContent component
