+++ src/hooks/useAuth.ts
@@ -1,43 +1,10 @@
-import { useState, useMemo } from "react";
-import { TSignInValues, TUseAuthReturn, TSignUpInputConstructor } from "../types";
-import { auth } from "../firebase/fbconfig";
+import { useMemo, useState } from "react";
 
-import { useContextSelector } from "use-context-selector";
-import { useNavigate } from "react-router-dom";
-import { AuthAPI } from "../store/AuthProvider";
+import { TSignInValues, TSignUpInputConstructor, TUseAuthReturn } from "../types";
 
 export const useAuth = (): TUseAuthReturn => {
-  const setLoading = useContextSelector(AuthAPI, (v) => v?.setLoading);
-  const navigate = useNavigate();
   const [signUpValues, setSignInValues] = useState<TSignInValues>({ email: "", password: "", passwordConfirm: "" });
   const [signUpError, setSignUpError] = useState<string>("");
-
-  const signup = (email: string, password: string) => {
-    return auth.createUserWithEmailAndPassword(email, password);
-  };
-
-  const handleRegister = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (signUpValues.password !== signUpValues.passwordConfirm) {
-      return setSignUpError("Passwords do not match");
-    }
-    if (!/^[a-zA-Z\s].*@.*$/.test(signUpValues.email)) {
-      return setSignUpError("Email is not correct");
-    }
-    if (signUpValues.password === "") {
-      return setSignUpError("Password field can not be empty");
-    } else {
-      try {
-        setSignUpError("");
-        setLoading?.(true);
-        await signup(signUpValues.email, signUpValues.password);
-        setTimeout(() => navigate("/"));
-      } catch (errors) {
-        setSignUpError("Failed to create an account");
-      }
-      setLoading?.(false);
-    }
-  };
 
   const SignUpInputConstructor: TSignUpInputConstructor = useMemo(
     () => [
@@ -81,5 +48,5 @@ export const useAuth = (): TUseAuthReturn => {
     [signUpValues]
   );
 
-  return { signUpValues, SignUpInputConstructor, handleRegister, signUpError };
+  return { signUpValues, SignUpInputConstructor, signUpError, setSignUpError };
 };
